Refactored files with datadog written in short, alphabetically ordered the DatadogConfig frozen class in types
