+++ src/apps/backend/modules/config/config_service.py
@@ -24,10 +24,10 @@ class ConfigService:
     def get_datadog_config() -> DatadogConfig:
         datadog_config = DictUtil.required_get_dict(input_dict=ConfigManager.config,key="DATADOG")
         return DatadogConfig(
-            api_key=DictUtil.required_get_str(input_dict=datadog_config, key="dd_api_key"),
-            host=DictUtil.required_get_str(input_dict=datadog_config,key="dd_site_name"),
-            app_name = DictUtil.required_get_str(input_dict=datadog_config,key="dd_app_name"),
-            dd_log_level = DictUtil.required_get_str(input_dict=datadog_config,key="dd_log_level")
+            api_key=DictUtil.required_get_str(input_dict=datadog_config, key="datadog_api_key"),
+            host=DictUtil.required_get_str(input_dict=datadog_config,key="datadog_site_name"),
+            app_name = DictUtil.required_get_str(input_dict=datadog_config,key="datadog_app_name"),
+            datadog_log_level = DictUtil.required_get_str(input_dict=datadog_config,key="datadog_log_level")
         )
 
     @staticmethod
