+++ src/apps/backend/modules/config/types.py
@@ -4,6 +4,6 @@ from dataclasses import dataclass
 @dataclass(frozen=True)
 class DatadogConfig:
     api_key: str
-    host: str
     app_name: str
-    dd_log_level: str
+    datadog_log_level: str
+    host: str
