+++ src/apps/backend/settings/os_envs.py
@@ -34,9 +34,9 @@ class OSSettings:
 
     DATADOG: Optional[dict[str, Optional[str]]] = field(
         default_factory=lambda: {
-            "dd_api_key": os.environ.get("DD_API_KEY"),
-            "dd_site_name": os.environ.get("DD_SITE"),
-            "dd_app_name": os.environ.get("DATADOG_APP_NAME"),
-            "dd_log_level": os.environ.get("DD_LOG_LEVEL")
+            "datadog_api_key": os.environ.get("DATADOG_API_KEY"),
+            "datadog_site_name": os.environ.get("DATADOG_SITE"),
+            "datadog_app_name": os.environ.get("DATADOG_APP_NAME"),
+            "datadog_log_level": os.environ.get("DATADOG_LOG_LEVEL")
         }
     )
